+++ tests/testthat/helper.R
@@ -1,6 +1,8 @@
 # nolint start
 library(mlr3)
 library(mlr3misc)
+library(mlr3pipelines)
+library(paradox)
 library(R6)
 
 lapply(list.files(system.file("testthat", package = "mlr3"), pattern = "^helper.*\\.[rR]", full.names = TRUE), source)